+++ src/main/java/com/gg/mlg/user/UserController.java
@@ -23,7 +23,7 @@ import java.util.List;
 import java.util.Map;
 
 @Controller
-@RequestMapping("/")
+@RequestMapping("/user")
 public class UserController {
 
     @Autowired
@@ -31,50 +31,8 @@ public class UserController {
     @Autowired
     private LolService lolservice;
 
-    /*@GetMapping("/home")
-    public String home(){
-        return "home";
-    }*/
 
-    @GetMapping("/home")
-    public String home(Principal principal, Model model){
-        model.addAttribute("data",0);
-        if(principal!=null){
-            UserEntity param= new UserEntity();
-            param.setUid(principal.getName());
-            param=service.selUser(param);
-            param.setUpw(null);
-            model.addAttribute("data",param);
-        }
-        return "home";
-    }
-    @PostMapping("/home")
-    public String index(String text, Model model) throws IOException {
-        List<String> hrefList = new ArrayList();
-        List<String> ImgList = new ArrayList();
-        List<String> titleList = new ArrayList();
-        List<String> writerList = new ArrayList();
-//        Map<String, Object> map = new HashMap<>();
-        String url = "https://www.google.com/search?q="+text+"+%EC%9C%A0%ED%8A%9C%EB%B8%8C&sxsrf=ALeKk01eqJ-QQtPbqLXUjxiIRUT3UneyOQ:1626873181915&source=lnms&tbm=vid&sa=X&ved=2ahUKEwi81ejwnvTxAhURxYsBHTYBBWgQ_AUoAXoECAEQAw&biw=1286&bih=788";
-        Document doc = Jsoup.connect(url).get();
-        Elements el = doc.select("a.rGhul");
-        String[] cnt = ((doc.select("div.ij69rd.UHe5G")).text().split(" ")); //조회수
-        for (Element e:el) {
-            hrefList.add(e.attr("href"));
-            Document doc2 = Jsoup.connect(e.attr("href")).get();
-            ImgList.add(doc2.select("link[itemprop=thumbnailUrl]").attr("href")); //썸네일
-            titleList.add(doc2.title().substring(0,doc2.title().length()-10)); //제목
-            writerList.add(doc2.select("link[itemprop=name]").attr("content")); //작성자
-        }
-        model.addAttribute("hrefList",hrefList);
-        model.addAttribute("ImgList",ImgList);
-        model.addAttribute("titleList",titleList);
-        model.addAttribute("writerList",writerList);
-        model.addAttribute("cntList",cnt);
-        return "temp/home";
-    }
-
-    @GetMapping("/user/login")
+    @GetMapping("login")
     public String login(HttpServletRequest request) {
         String refer=request.getHeader("Referer");
         if(refer.contains("/user/login")){
@@ -84,7 +42,7 @@ public class UserController {
         return "user/login";
     }
 
-    @GetMapping("/user/join")
+    @GetMapping("join")
     public String join() {
         return "user/join";
     }
@@ -96,13 +54,13 @@ public class UserController {
     }
 
     @ResponseBody
-    @PostMapping("/user/follow")
+    @PostMapping("follow")
     public int follow(@RequestBody UserEntity param){
         System.out.println(param);
         return service.indelfollow(param);
     }
 
-    @GetMapping("/user/profile")
+    @GetMapping("profile")
     public String profile(@Parameter UserEntity param, Model model) {
         UserEntity result=service.selUser(param);
         model.addAttribute("pud", result);
